Make sure to use an absolute path for the SwiftLint configuration
When SwiftLint is being run, the working directory is _not_ the project's
root directory, resulting in the SwiftLint configuration file not being found.
By calculating the absolute path from the relative one that was provided,
it is made sure that the configuration file is found regardless of the current
working directory.